refactor: extract configSchema into spec.ts for each connector
Move Zod config schemas from index.ts into dedicated spec.ts files and
rename `spec` → `configSchema` across all three connectors:

- packages/source-stripe/src/spec.ts
- packages/destination-postgres/src/spec.ts
- packages/destination-google-sheets/src/spec.ts

Each index.ts re-exports configSchema + Config type. bin.ts files import
from spec.ts directly. Conformance tests updated to check for configSchema.

Committed-By-Agent: claude
